fix(vscode): tail-match builder IDs in openBuilderByRoleOrId
Fixes 'Codev: No active terminal for 153' when clicking a builder row
in the sidebar. The sidebar passes OverviewBuilder.id (e.g. '153' for
SPIR builders, parsed from the .builders/spir-153-*/ dir name in
overview.ts), but Tower's runtime Builder.id is the canonical
'builder-spir-153'. The strict-equality lookup never matched.

Lifts resolveAgentName + stripLeadingZeros from
packages/codev/src/agent-farm/utils/agent-names.ts into a new
@cluesmith/codev-core/agent-names subpath, generic over { id: string }
so it works for both the agent-farm internal Builder shape and the
public codev-types Builder. The agent-farm utility re-exports from
core, so existing callsites are unaffected. TerminalManager imports
resolveAgentName and uses its tail-match fallback (target '153'
matches builderId 'builder-spir-153' because it ends with '-153').
